+++ internal/blobcache/blobcache.go
@@ -1,4 +1,4 @@
-package fuse
+package blobcache
 
 import (
 	"sync"
@@ -11,11 +11,11 @@ import (
 
 // Crude estimate of the overhead per blob: a SHA-256, a linked list node
 // and some pointers. See comment in blobCache.add.
-const cacheOverhead = len(restic.ID{}) + 64
+const CacheOverhead = len(restic.ID{}) + 64
 
 // A blobCache is a fixed-size cache of blob contents.
 // It is safe for concurrent access.
-type blobCache struct {
+type BlobCache struct {
 	mu sync.Mutex
 	c  *simplelru.LRU
 
@@ -23,16 +23,16 @@ type blobCache struct {
 }
 
 // Construct a blob cache that stores at most size bytes worth of blobs.
-func newBlobCache(size int) *blobCache {
-	c := &blobCache{
+func New(size int) *BlobCache {
+	c := &BlobCache{
 		free: size,
 		size: size,
 	}
 
 	// NewLRU wants us to specify some max. number of entries, else it errors.
-	// The actual maximum will be smaller than size/cacheOverhead, because we
+	// The actual maximum will be smaller than size/CacheOverhead, because we
 	// evict entries (RemoveOldest in add) to maintain our size bound.
-	maxEntries := size / cacheOverhead
+	maxEntries := size / CacheOverhead
 	lru, err := simplelru.NewLRU(maxEntries, c.evict)
 	if err != nil {
 		panic(err) // Can only be maxEntries <= 0.
@@ -42,10 +42,10 @@ func newBlobCache(size int) *blobCache {
 	return c
 }
 
-func (c *blobCache) add(id restic.ID, blob []byte) {
+func (c *BlobCache) Add(id restic.ID, blob []byte) {
 	debug.Log("blobCache: add %v", id)
 
-	size := len(blob) + cacheOverhead
+	size := len(blob) + CacheOverhead
 	if size > c.size {
 		return
 	}
@@ -59,7 +59,7 @@ func (c *blobCache) add(id restic.ID, blob []byte) {
 		return
 	}
 
-	// This loop takes at most min(maxEntries, maxchunksize/cacheOverhead)
+	// This loop takes at most min(maxEntries, maxchunksize/CacheOverhead)
 	// iterations.
 	for size > c.free {
 		c.c.RemoveOldest()
@@ -69,19 +69,19 @@ func (c *blobCache) add(id restic.ID, blob []byte) {
 	c.free -= size
 }
 
-func (c *blobCache) get(id restic.ID) ([]byte, bool) {
+func (c *BlobCache) Get(id restic.ID) ([]byte, bool) {
 	c.mu.Lock()
 	value, ok := c.c.Get(id)
 	c.mu.Unlock()
 
-	debug.Log("blobCache: get %v, hit %v", id, ok)
+	debug.Log("BlobCache: get %v, hit %v", id, ok)
 
 	blob, ok := value.([]byte)
 	return blob, ok
 }
 
-func (c *blobCache) evict(key, value interface{}) {
+func (c *BlobCache) evict(key, value interface{}) {
 	blob := value.([]byte)
-	debug.Log("blobCache: evict %v, %d bytes", key, len(blob))
-	c.free += len(blob) + cacheOverhead
+	debug.Log("BlobCache: evict %v, %d bytes", key, len(blob))
+	c.free += len(blob) + CacheOverhead
 }
